@@ -4,7 +4,7 @@ version: 1.1.1
 homepage: https://github.com/sefidgaran/signalr_client
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 
