@@ -8,41 +8,36 @@ class SignalRHttpRequest {
   // Properties
 
   /// The HTTP method to use for the request.
-  String method;
+  String? method;
 
   /// The URL for the request.
-  String url;
+  String? url;
 
   ///The body content for the request. May be a string or an ArrayBuffer (for binary data).
-  Object content;
+  Object? content;
 
   ///An object describing headers to apply to the request.
-  MessageHeaders headers;
+  MessageHeaders? headers;
 
   // /** The XMLHttpRequestResponseType to apply to the request. */
   // responseType?: XMLHttpRequestResponseType;
 
   /// An AbortSignal that can be monitored for cancellation.
-  IAbortSignal abortSignal;
+  IAbortSignal? abortSignal;
 
   ///The time to wait for the request to complete before throwing a TimeoutError. Measured in milliseconds.
-  int timeout;
+  int? timeout;
 
   // Methods
 
-  SignalRHttpRequest(
-      {String method,
-      String url,
-      Object content,
-      MessageHeaders headers,
-      IAbortSignal abortSignal,
-      int timeout})
-      : this.method = method,
-        this.url = url,
-        this.content = content,
-        this.headers = headers,
-        this.abortSignal = abortSignal,
-        this.timeout = timeout;
+  SignalRHttpRequest({
+    this.url,
+    this.method,
+    this.content,
+    this.headers,
+    this.abortSignal,
+    this.timeout,
+  });
 }
 
 /// Represents an HTTP response.
@@ -53,10 +48,10 @@ class SignalRHttpResponse {
   final int statusCode;
 
   /// The status message of the response
-  final String statusText;
+  final String? statusText;
 
   /// May be a string (json) or an Uint8List (binary)
-  final Object content;
+  final Object? content;
 
   //Methods
 
@@ -66,7 +61,8 @@ class SignalRHttpResponse {
   /// statusText: The status message of the response.
   /// content: The content of the response
   ///
-  SignalRHttpResponse(int statusCode, {String statusText = '', Object content})
+  SignalRHttpResponse(int statusCode,
+      {String? statusText = '', Object? content})
       : this.statusCode = statusCode,
         this.statusText = statusText,
         this.content = content;
@@ -83,8 +79,8 @@ abstract class SignalRHttpClient {
   /// HttpRequest options Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future<HttpResponse> that resolves with an HttpResponse describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> get(String url, {SignalRHttpRequest options}) {
-    options.method = 'GET';
+  Future<SignalRHttpResponse> get(String url, {SignalRHttpRequest? options}) {
+    options!.method = 'GET';
     options.url = url;
     return send(options);
   }
@@ -95,8 +91,8 @@ abstract class SignalRHttpClient {
   /// options: Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future that resolves with an describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> post(String url, {SignalRHttpRequest options}) {
-    options.method = 'POST';
+  Future<SignalRHttpResponse> post(String url, {SignalRHttpRequest? options}) {
+    options!.method = 'POST';
     options.url = url;
     return send(options);
   }
@@ -107,8 +103,9 @@ abstract class SignalRHttpClient {
   /// Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future that resolves with an describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> delete(String url, {SignalRHttpRequest options}) {
-    options.method = 'DELETE';
+  Future<SignalRHttpResponse> delete(String url,
+      {SignalRHttpRequest? options}) {
+    options!.method = 'DELETE';
     options.url = url;
     return send(options);
   }
