@@ -25,95 +25,87 @@ enum ConnectionState {
 
 class NegotiateResponse {
   // Properties
-  String connectionId;
-  String connectionToken;
-  int negotiateVersion;
-  List<AvailableTransport> availableTransports;
-  final String url;
-  final String accessToken;
-  final String error;
+  final List<AvailableTransport> availableTransports;
+  String? url;
+  final String? accessToken;
+  final String? error;
 
-  bool get hasConnectionId => !isStringEmpty(connectionId);
+  String? connectionId;
+  String? connectionToken;
+  int? negotiateVersion;
 
-  bool get hasConnectionTokenId => !isStringEmpty(connectionToken);
+  bool get hasConnectionId => connectionId.isNotNullOrEmpty;
 
-  bool get hasNegotiateVersion => !isIntEmpty(negotiateVersion);
+  bool get hasConnectionTokenId => connectionToken.isNotNullOrEmpty;
+
+  bool get hasNegotiateVersion => negotiateVersion == null;
 
   bool get isConnectionResponse =>
-      hasConnectionId && !isListEmpty(availableTransports);
+      hasConnectionId && availableTransports.isEmpty;
 
-  bool get isRedirectResponse => !isStringEmpty(url);
+  bool get isRedirectResponse => url.isNotNullOrEmpty;
 
-  bool get isErrorResponse => !isStringEmpty(error);
+  bool get isErrorResponse => error.isNotNullOrEmpty;
 
-  bool get hasAccessToken => !isStringEmpty(accessToken);
+  bool get hasAccessToken => accessToken.isNotNullOrEmpty;
 
   // Methods
 
-  NegotiateResponse(
-      this.connectionId,
-      this.connectionToken,
-      this.negotiateVersion,
-      this.availableTransports,
-      this.url,
-      this.accessToken,
-      this.error);
+  NegotiateResponse({
+    required this.url,
+    this.connectionId,
+    this.connectionToken,
+    this.negotiateVersion,
+    this.availableTransports = const [],
+    this.accessToken,
+    this.error,
+  });
 
-  NegotiateResponse.fromJson(Map<String, dynamic> json)
-      : assert(json != null),
-        this.connectionId = json['connectionId'],
-        this.connectionToken = json['connectionToken'],
-        this.negotiateVersion = json['negotiateVersion'],
-        this.url = json['url'],
-        this.accessToken = json['accessToken'],
-        this.error = json['error'] {
-    availableTransports = [];
-    final List<dynamic> transports = json['availableTransports'];
-    if (transports == null) {
-      return;
-    }
-
-    for (var i = 0; i < transports.length; i++) {
-      availableTransports.add(AvailableTransport.fromJson(transports[i]));
-    }
-  }
+  factory NegotiateResponse.fromJson(Map<String, dynamic> json) =>
+      NegotiateResponse(
+          connectionId: json['connectionId'],
+          connectionToken: json['connectionToken'],
+          negotiateVersion: json['negotiateVersion'],
+          availableTransports:
+              ((json['availableTransports'] as List<dynamic>?) ?? [])
+                  .map((format) => AvailableTransport.fromJson(format))
+                  .toList(),
+          url: json['url'],
+          accessToken: json['accessToken'],
+          error: json['error']);
 }
 
 class AvailableTransport {
   // Properties
 
   HttpTransportType transport;
-  List<TransferFormat> transferFormats;
+  late List<TransferFormat> transferFormats;
 
   // Methods
 
   AvailableTransport(this.transport, this.transferFormats);
 
-  AvailableTransport.fromJson(Map<String, dynamic> json) {
-    transferFormats = [];
+  factory AvailableTransport.fromJson(Map<String, dynamic> json) {
+    final transport = httpTransportTypeFromString(json['transport']);
+    List<dynamic>? formats = json['transferFormats'];
 
-    if (json == null) {
-      return;
-    }
+    if (formats == null)
+      throw Exception(
+          "No formats found for transport in the provided json. Make sure that the provided json has the"
+          "transferFormats");
 
-    transport = httpTransportTypeFromString(json['transport']);
-    List<dynamic> formats = json['transferFormats'];
-    if (formats == null) {
-      return;
-    }
-    for (var i = 0; i < formats.length; i++) {
-      transferFormats.add(getTransferFormatFromString(formats[i]));
-    }
+    return AvailableTransport(transport,
+        formats.map((format) => getTransferFormatFromString(format)).toList());
   }
 }
 
 class TransportSendQueue {
   List<Object> _buffer = [];
-  Completer _sendBufferedData;
+  late Completer _sendBufferedData;
   bool _executing = true;
-  Completer _transportResult;
-  Future<void> _sendLoopPromise;
-  final ITransport transport;
+  Completer? _transportResult;
+  Future<void>? _sendLoopPromise;
+  final ITransport? transport;
 
   TransportSendQueue(this.transport) {
     _sendBufferedData = Completer();
@@ -128,10 +120,10 @@ class TransportSendQueue {
       _transportResult = Completer();
     }
 
-    return _transportResult.future;
+    return _transportResult!.future;
   }
 
-  Future<void> stop() {
+  Future<void>? stop() {
     _executing = false;
     if (!_sendBufferedData.isCompleted) _sendBufferedData.complete();
     return _sendLoopPromise;
@@ -158,8 +150,8 @@ class TransportSendQueue {
 
       if (!_executing) {
         if (_transportResult != null) {
-          if (!_transportResult.isCompleted) {
-            _transportResult.completeError('Connection stopped.');
+          if (!_transportResult!.isCompleted) {
+            _transportResult!.completeError('Connection stopped.');
           }
         }
 
@@ -168,17 +160,17 @@ class TransportSendQueue {
 
       _sendBufferedData = Completer();
 
-      var transportResult = _transportResult;
+      var transportResult = _transportResult!;
       _transportResult = null;
 
       var data = _buffer[0] is String
           ? _buffer.join("")
-          : TransportSendQueue.concatBuffers(_buffer);
+          : TransportSendQueue.concatBuffers(_buffer as List<Uint8List>);
 
       _buffer.length = 0;
 
       try {
-        await this.transport.send(data);
+        await this.transport!.send(data);
         if (!transportResult.isCompleted) transportResult.complete();
       } catch (error) {
         if (!transportResult.isCompleted) transportResult.completeError(error);
@@ -204,49 +196,47 @@ class HttpConnection implements IConnection {
   static final maxRedirects = 100;
   static final maxRequestTimeoutMilliseconds = 2000;
 
-  ConnectionState _connectionState;
+  ConnectionState? _connectionState = ConnectionState.Disconnected;
 
   // connectionStarted is tracked independently from connectionState, so we can check if the
   // connection ever did successfully transition from connecting to connected before disconnecting.
-  bool _connectionStarted;
-  SignalRHttpClient _httpClient;
-  final Logger _logger;
-  HttpConnectionOptions _options;
-  ITransport _transport;
-  Future<void> _startInternalPromise;
-  Future<void> _stopPromise;
-  Completer _stopPromiseCompleter;
-  Exception _stopError;
-  AccessTokenFactory _accessTokenFactory;
-  TransportSendQueue _sendQueue;
+  bool _connectionStarted = false;
+  final SignalRHttpClient _httpClient;
+  final Logger? _logger;
+  final HttpConnectionOptions _options;
+  ITransport? _transport;
+  Future<void>? _startInternalPromise;
 
-  ConnectionFeatures features;
+  late Future<void> _stopPromise;
+  late Completer _stopPromiseCompleter;
+  Exception? _stopError;
+  AccessTokenFactory? _accessTokenFactory;
+  TransportSendQueue? _sendQueue;
+
+  ConnectionFeatures? features;
   String baseUrl;
-  String connectionId;
+  String? connectionId;
 
   @override
-  OnReceive onreceive;
+  OnReceive? onReceive;
 
   @override
-  OnClose onclose;
+  OnClose? onClose;
 
   int _negotiateVersion = 1;
 
   // Methods
 
-  HttpConnection(String url, {HttpConnectionOptions options})
-      : assert(url != null),
-        _logger = options?.logger {
-    baseUrl = url;
-
-    _options = options ?? HttpConnectionOptions();
-    _httpClient = options.httpClient ?? WebSupportingHttpClient(_logger);
-    _connectionState = ConnectionState.Disconnected;
-    _connectionStarted = false;
-  }
+  HttpConnection(String url,
+      {HttpConnectionOptions options = const HttpConnectionOptions()})
+      : baseUrl = url,
+        _options = options,
+        _httpClient =
+            options.httpClient ?? WebSupportingHttpClient(options.logger!),
+        _logger = options.logger;
 
   @override
-  Future<void> start({TransferFormat transferFormat}) async {
+  Future<void> start({TransferFormat? transferFormat}) async {
     transferFormat = transferFormat ?? TransferFormat.Binary;
 
     _logger
@@ -296,15 +286,15 @@ class HttpConnection implements IConnection {
     }
 
     // Transport will not be null if state is connected
-    return _sendQueue.send(data);
+    return _sendQueue!.send(data);
   }
 
   @override
-  Future<void> stop({Exception error}) async {
+  Future<void> stop({Exception? error}) async {
     if (_connectionState == ConnectionState.Disconnected) {
       _logger?.finer(
           "Call to HttpConnection.stop($error) ignored because the connection is already in the disconnected state.");
-      return Future.value();
+      return;
     }
 
     if (_connectionState == ConnectionState.Disconnecting) {
@@ -324,7 +314,7 @@ class HttpConnection implements IConnection {
     await _stopPromise;
   }
 
-  Future<void> _stopInternal({Exception error}) async {
+  Future<void> _stopInternal({Exception? error}) async {
     // Set error as soon as possible otherwise there is a race between
     // the transport closing and providing an error and the error from a close message
     // We would prefer the close message error.
@@ -341,7 +331,7 @@ class HttpConnection implements IConnection {
     // stop was called during start() and start() failed.
     if (_transport != null) {
       try {
-        await _transport.stop();
+        await _transport!.stop();
       } catch (e) {
         _logger?.severe("HttpConnection.transport.stop() threw error '$e'.");
         _stopConnection();
@@ -394,8 +384,9 @@ class HttpConnection implements IConnection {
           //     throw GeneralError("Detected a connection attempt to an ASP.NET SignalR Server. This client only supports connecting to an ASP.NET Core SignalR Server. See https://aka.ms/signalr-core-differences for details.");
           // }
 
-          if (negotiateResponse.isRedirectResponse) {
-            url = negotiateResponse.url;
+          if (negotiateResponse.isRedirectResponse &&
+              negotiateResponse.url.isNotNullOrEmpty) {
+            url = negotiateResponse.url!;
           }
 
           if (negotiateResponse.hasAccessToken) {
@@ -422,7 +413,7 @@ class HttpConnection implements IConnection {
         if (features == null) {
           features = ConnectionFeatures(true);
         } else {
-          features.inherentKeepAlive = true;
+          features!.inherentKeepAlive = true;
         }
       }
 
@@ -447,10 +438,8 @@ class HttpConnection implements IConnection {
   Future<NegotiateResponse> _getNegotiationResponse(String url) async {
     MessageHeaders headers = MessageHeaders();
     if (_accessTokenFactory != null) {
-      final token = await _accessTokenFactory();
-      if (token != null) {
-        headers.setHeaderValue("Authorization", "Bearer $token");
-      }
+      final token = await _accessTokenFactory!();
+      headers["Authorization"] = "Bearer $token";
     }
 
     final negotiateUrl = _resolveNegotiateUrl(url);
@@ -469,13 +458,13 @@ class HttpConnection implements IConnection {
 
       if (!(response.content is String)) {
         return Future.error(
-            GeneralError("Negotation response content must be a json."));
+            GeneralError("Negotiation response content must be a json."));
       }
 
       var negotiateResponse =
           NegotiateResponse.fromJson(json.decode(response.content as String));
       if (negotiateResponse.negotiateVersion == null ||
-          negotiateResponse.negotiateVersion < 1) {
+          negotiateResponse.negotiateVersion! < 1) {
         // Negotiate version 0 doesn't use connectionToken
         // So we set it equal to connectionId so all our logic can use connectionToken without being aware of the negotiate version
         negotiateResponse.connectionToken = negotiateResponse.connectionId;
@@ -488,7 +477,7 @@ class HttpConnection implements IConnection {
     }
   }
 
-  String _createConnectUrl(String url, String connectionToken) {
+  String _createConnectUrl(String url, [String? connectionToken]) {
     if (connectionToken == null) {
       return url;
     }
@@ -498,14 +487,14 @@ class HttpConnection implements IConnection {
 
   Future<void> _createTransport(
       String url,
-      Object requestedTransport,
+      Object? requestedTransport,
       NegotiateResponse negotiateResponse,
       TransferFormat requestedTransferFormat) async {
     var connectUrl = _createConnectUrl(url, negotiateResponse.connectionToken);
     if (_isITransport(requestedTransport)) {
       _logger?.finer(
           "Connection was provided an instance of ITransport, using that directly.");
-      _transport = requestedTransport;
+      _transport = requestedTransport as ITransport?;
       await _startTransport(connectUrl, requestedTransferFormat);
 
       connectionId = negotiateResponse.connectionId;
@@ -514,13 +503,13 @@ class HttpConnection implements IConnection {
 
     final List<Object> transportExceptions = [];
     final transports = negotiateResponse.availableTransports ?? [];
-    NegotiateResponse negotiate = negotiateResponse;
+    NegotiateResponse? negotiate = negotiateResponse;
     for (var endpoint in transports) {
       _connectionState = ConnectionState.Connecting;
 
       try {
-        _transport = _resolveTransport(
-            endpoint, requestedTransport, requestedTransferFormat);
+        _transport = _resolveTransport(endpoint,
+            requestedTransport as HttpTransportType?, requestedTransferFormat);
       } catch (e) {
         transportExceptions.add("${endpoint.transport} failed: $e");
         continue;
@@ -528,7 +517,7 @@ class HttpConnection implements IConnection {
 
       if (negotiate == null) {
         try {
-          negotiate = await _getNegotiationResponse(url);
+          negotiate = await _getNegotiationResponse(url!);
         } catch (ex) {
           return Future.error(ex);
         }
@@ -579,51 +568,45 @@ class HttpConnection implements IConnection {
   }
 
   Future<void> _startTransport(String url, TransferFormat transferFormat) {
-    _transport.onReceive = onreceive;
-    _transport.onClose = _stopConnection;
-    return _transport.connect(url, transferFormat);
+    _transport!.onReceive = onReceive;
+    _transport!.onClose = _stopConnection;
+    return _transport!.connect(url, transferFormat);
   }
 
   ITransport _resolveTransport(
       AvailableTransport endpoint,
-      HttpTransportType requestedTransport,
+      HttpTransportType? requestedTransport,
       TransferFormat requestedTransferFormat) {
     final transport = endpoint.transport;
-    if (transport == null) {
-      _logger?.finer(
-          "Skipping transport '${endpoint.transport}' because it is not supported by this client.");
-      throw GeneralError(
-          "Skipping transport '${endpoint.transport}' because it is not supported by this client.");
-    } else {
-      if (transportMatches(requestedTransport, transport)) {
-        final transferFormats = endpoint.transferFormats;
-        if (transferFormats.indexOf(requestedTransferFormat) >= 0) {
-          _logger?.finer("Selecting transport '${transport.toString()}'.");
-          try {
-            return _constructTransport(transport);
-          } catch (ex) {
-            return ex;
-          }
-        } else {
-          _logger?.finer(
-              "Skipping transport '$transport' because it does not support the requested transfer format '$requestedTransferFormat'.");
-          throw GeneralError(
-              "Skipping transport '$transport' because it does not support the requested transfer format '$requestedTransferFormat'.");
+
+    if (transportMatches(requestedTransport, transport)) {
+      final transferFormats = endpoint.transferFormats;
+      if (transferFormats.indexOf(requestedTransferFormat) >= 0) {
+        _logger?.finer("Selecting transport '${transport.toString()}'.");
+        try {
+          return _constructTransport(transport);
+        } catch (ex) {
+          rethrow;
         }
       } else {
         _logger?.finer(
-            "Skipping transport '$transport' because it was disabled by the client.");
+            "Skipping transport '$transport' because it does not support the requested transfer format '$requestedTransferFormat'.");
         throw GeneralError(
-            "Skipping transport '$transport' because it was disabled by the client.");
+            "Skipping transport '$transport' because it does not support the requested transfer format '$requestedTransferFormat'.");
       }
+    } else {
+      _logger?.finer(
+          "Skipping transport '$transport' because it was disabled by the client.");
+      throw GeneralError(
+          "Skipping transport '$transport' because it was disabled by the client.");
     }
   }
 
-  bool _isITransport(Object transport) {
+  bool _isITransport(Object? transport) {
     return transport is ITransport;
   }
 
-  void _stopConnection({Exception error}) {
+  void _stopConnection({Exception? error}) {
     _logger?.finer(
         "HttpConnection.stopConnection(${error ?? "Unknown"}) called while in state $_connectionState.");
 
@@ -659,7 +642,7 @@ class HttpConnection implements IConnection {
     }
 
     if (_sendQueue != null) {
-      _sendQueue.stop().catchError((e) {
+      _sendQueue!.stop()!.catchError((e) {
         _logger?.severe("TransportSendQueue.stop() threw error '$e'.");
       });
       _sendQueue = null;
@@ -672,8 +655,8 @@ class HttpConnection implements IConnection {
       _connectionStarted = false;
 
       try {
-        if (onclose != null) {
-          onclose(error: error);
+        if (onClose != null) {
+          onClose!(error: error);
         }
       } catch (e) {
         _logger?.severe("HttpConnection.onclose($error) threw error '$e'.");
@@ -697,8 +680,8 @@ class HttpConnection implements IConnection {
     return negotiateUrl;
   }
 
-  static bool transportMatches(
-      HttpTransportType requestedTransport, HttpTransportType actualTransport) {
+  static bool transportMatches(HttpTransportType? requestedTransport,
+      HttpTransportType actualTransport) {
     return (requestedTransport == null) ||
         (actualTransport == requestedTransport);
   }
