@@ -7,39 +7,39 @@ class HttpConnectionOptions {
   // Properties
 
   /// An SignalRHttpClient that will be used to make HTTP requests.
-  SignalRHttpClient httpClient;
+  final SignalRHttpClient? httpClient;
 
   /// An HttpTransportType or ITransport value specifying the transport to use for the connection
   /// If transport is null and the server supports all transport protocols than HttpTransportType.WebSockets is used.
-  Object transport;
+  final Object? transport;
 
   /// Configures the logger used for logging.
   ///
   /// Provide an Logger instance, and log messages will be logged via that instance
   ///
-  Logger logger;
+  final Logger? logger;
 
   /// A function that provides an access token required for HTTP Bearer authentication.
-  AccessTokenFactory accessTokenFactory;
+  final AccessTokenFactory? accessTokenFactory;
 
   /// A boolean indicating if message content should be logged.
   ///
   /// Message content can contain sensitive user data, so this is disabled by default.
   ///
-  bool logMessageContent;
+  final bool logMessageContent;
 
   /// A boolean indicating if negotiation should be skipped.
   ///
   /// Negotiation can only be skipped when the IHttpConnectionOptions.transport property is set to 'HttpTransportType.WebSockets'.
   ///
-  bool skipNegotiation;
+  final bool skipNegotiation;
 
   // Methods
-  HttpConnectionOptions(
-      {SignalRHttpClient httpClient,
-      Object transport,
-      Logger logger,
-      AccessTokenFactory accessTokenFactory,
+  const HttpConnectionOptions(
+      {SignalRHttpClient? httpClient,
+      Object? transport,
+      Logger? logger,
+      AccessTokenFactory? accessTokenFactory,
       bool logMessageContent = false,
       bool skipNegotiation = false})
       : this.httpClient = httpClient,
