@@ -4,24 +4,24 @@ import '../transport/itransport.dart';
 
 class ConnectionFeatures {
   // Properties
-  bool inherentKeepAlive;
+  bool? inherentKeepAlive;
 
   // Methods
   ConnectionFeatures(this.inherentKeepAlive);
 }
 
 abstract class IConnection {
-  ConnectionFeatures features;
-  String connectionId;
+  ConnectionFeatures? features;
+  String? connectionId;
 
-  String baseUrl;
+  late String baseUrl;
 
-  Future<void> start({TransferFormat transferFormat});
+  Future<void> start({TransferFormat? transferFormat});
   Future<void> send(Object data);
-  Future<void> stop({Exception error});
+  Future<void> stop({Exception? error});
 
-  OnReceive onreceive;
-  OnClose onclose;
+  OnReceive? onReceive;
+  OnClose? onClose;
 
   IConnection() : features = ConnectionFeatures(null);
 }
