@@ -1,9 +1,9 @@
 import 'iretry_policy.dart';
 
 class DefaultRetryPolicy implements IRetryPolicy {
-  List<int> _retryDelays;
+  late List<int?> _retryDelays;
 
-  static const List<int> DEFAULT_RETRY_DELAYS_IN_MILLISECONDS = [
+  static const List<int?> DEFAULT_RETRY_DELAYS_IN_MILLISECONDS = [
     0,
     2000,
     10000,
@@ -11,14 +11,14 @@ class DefaultRetryPolicy implements IRetryPolicy {
     null
   ];
 
-  DefaultRetryPolicy({List<int> retryDelays}) {
+  DefaultRetryPolicy({List<int>? retryDelays}) {
     _retryDelays = retryDelays != null
         ? [...retryDelays, null]
         : DEFAULT_RETRY_DELAYS_IN_MILLISECONDS;
   }
 
   @override
-  int nextRetryDelayInMilliseconds(RetryContext retryContext) {
+  int? nextRetryDelayInMilliseconds(RetryContext retryContext) {
     return _retryDelays[retryContext.previousRetryCount];
   }
 }
