@@ -32,14 +32,14 @@ class JsonHubProtocol implements IHubProtocol {
   /// A logger that will be used to log messages that occur during parsing.
   ///
   @override
-  List<HubMessageBase> parseMessages(Object input, Logger logger) {
+  List<HubMessageBase> parseMessages(Object input, Logger? logger) {
     // Only JsonContent is allowed.
     if (!(input is String)) {
       throw new GeneralError(
           "Invalid input for JSON hub protocol. Expected a string.");
     }
 
-    final jsonInput = input as String;
+    final jsonInput = input;
     final List<HubMessageBase> hubMessages = [];
 
     if (input == null) {
@@ -80,12 +80,12 @@ class JsonHubProtocol implements IHubProtocol {
     return hubMessages;
   }
 
-  static MessageType _getMessageTypeFromJson(Map<String, dynamic> json) {
+  static MessageType? _getMessageTypeFromJson(Map<String, dynamic> json) {
     return parseMessageTypeFromString(json["type"]);
   }
 
-  static MessageHeaders createMessageHeadersFromJson(
-      Map<String, dynamic> jsonData) {
+  static MessageHeaders? createMessageHeadersFromJson(
+      Map<String, dynamic>? jsonData) {
     if (jsonData != null) {
       throw GeneralError("ToDo");
     }
@@ -94,10 +94,15 @@ class JsonHubProtocol implements IHubProtocol {
 
   static InvocationMessage _getInvocationMessageFromJson(
       Map<String, dynamic> jsonData) {
-    final MessageHeaders headers =
-        createMessageHeadersFromJson(jsonData["headers"]);
-    final message = InvocationMessage(jsonData["target"], jsonData["arguments"],
-        jsonData["streamIds"], headers, jsonData["invocationId"]);
+    final message = InvocationMessage(
+      target: jsonData["target"],
+      streamIds: jsonData["streamIds"] == null
+          ? []
+          : jsonData["streamIds"] as List<String>,
+      arguments: List<Object>.from(jsonData["arguments"]),
+      headers: createMessageHeadersFromJson(jsonData["headers"]),
+      invocationId: jsonData["invocationId"],
+    );
 
     _assertNotEmptyString(
         message.target, "Invalid payload for Invocation message.");
@@ -111,7 +116,7 @@ class JsonHubProtocol implements IHubProtocol {
 
   static StreamItemMessage _getStreamItemMessageFromJson(
       Map<String, dynamic> jsonData) {
-    final MessageHeaders headers =
+    final MessageHeaders? headers =
         createMessageHeadersFromJson(jsonData["headers"]);
     final message =
         StreamItemMessage(jsonData["item"], headers, jsonData["invocationId"]);
@@ -126,7 +131,7 @@ class JsonHubProtocol implements IHubProtocol {
 
   static CompletionMessage _getCompletionMessageFromJson(
       Map<String, dynamic> jsonData) {
-    final MessageHeaders headers =
+    final MessageHeaders? headers =
         createMessageHeadersFromJson(jsonData["headers"]);
     final message = CompletionMessage(jsonData["error"], jsonData["result"],
         headers, jsonData["invocationId"]);
@@ -173,7 +178,7 @@ class JsonHubProtocol implements IHubProtocol {
       throw GeneralError("Cannot encode message of type '${message.typ}'.");
     }
 
-    final messageType = (message as HubMessageBase).type.index;
+    final messageType = message.type.index;
 
     if (message is InvocationMessage) {
       return {
@@ -231,8 +236,8 @@ class JsonHubProtocol implements IHubProtocol {
     throw GeneralError("Converting '${message.type}' is not implemented.");
   }
 
-  static void _assertNotEmptyString(String value, String errorMessage) {
-    if (isStringEmpty(value)) {
+  static void _assertNotEmptyString(String? value, String errorMessage) {
+    if (value.isNullOrEmpty) {
       throw InvalidPayloadException(errorMessage);
     }
   }
